automatically backed up by learn

+++ lib/create.sql
@@ -17,7 +17,7 @@ CREATE TABLE pledges (
   id INTEGER PRIMARY KEY,
   user_id INTEGER,
   project_id INTEGER,
-  amount REAL
+  amount INTEGER
 );
